Move disassembler to separate file. Add error reporting
