+++ language/compiler/bytecode_source_map/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod disassembler;
 pub mod mapping;
 pub mod marking;
 pub mod source_map;
